wip navbar and other components
