+++ client/package.json
@@ -7,6 +7,7 @@
     "axios": "^0.18.0",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
+    "react-materialize": "^2.4.8",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.1"
   },
