--- client/src/components/Jumbotron/Jumbotron.css
@@ -1,4 +1,0 @@
-.jumbotron {
-  min-height: 300px;
-  padding-top: 100px;
-}
