--- client/src/components/Jumbotron/Jumbotron.js
@@ -1,11 +1,0 @@
-import React from "react";
-import "./Jumbotron.css";
-
-const Jumbotron = () => (
-  <div className="jumbotron text-center">
-    <h1 className= "display-4">FreshTomatoes</h1>
-    <p class="lead">Freshly picked movies just for you..</p>
-  </div>
-);
-
-export default Jumbotron;
