+++ client/src/components/Navbar-old/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Navbar";
