+++ client/src/components/styles/Navbar.css
@@ -1,0 +1,11 @@
+a, href {
+  color: black !important;
+}
+
+input {
+  color: black !important;
+}
+
+.material-icons {
+  color: black !important;
+}
